Read images from folder, delete afterwards
This version of camera-feed reads image files from folder "images/",
formulates a POST request to the Tensorflow model, and shows the result
locally in stdout.

Compared with previous version, this uses OpenCV to read the image file.
This seems to create the right data structure for the model.

It does not report the results anywhere sensible.

Warning: it deletes the images, so it is a good idea to do a backup